refactor: extract reusable serializer logic into dedicated module
- Moved built-in and custom serializer definitions to a new `Serializer` module for better reusability and maintainability.
- Replaced inline serializer logic within `KtConfigSymbolProcessor` with references to `Serializer`.
- Simplified type-based serializer resolution using `Serializer.findSerializer`.
- Removed redundant and deprecated serializer methods, including `getComment` and `getPathName`.
- Updated `build.gradle.kts` to use the new localization for server JAR URLs.
