+++ ksp/src/main/kotlin/dev/s7a/ktconfig/ksp/KSPHelpers.kt
@@ -1,0 +1,31 @@
+package dev.s7a.ktconfig.ksp
+
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSFile
+
+/**
+ * Gets the full name of a class by traversing its parent hierarchy.
+ * For nested classes, returns a list of class names from outermost to innermost.
+ * For top-level classes, returns a single-element list with the class name.
+ *
+ * @param declaration The class declaration to get the full name for
+ * @return List of class names representing the full hierarchy
+ */
+fun getFullName(declaration: KSClassDeclaration): List<String> =
+    if (declaration.parent is KSFile) {
+        listOf(declaration.simpleName.asString())
+    } else {
+        getFullName(declaration.parent as KSClassDeclaration) + declaration.simpleName.asString()
+    }
+
+/**
+ * Generates a loader class name for the given class declaration.
+ * Combines the full class name parts with underscores and appends "Loader".
+ *
+ * @param declaration The class declaration to generate a loader name for
+ * @return The generated loader class name
+ */
+fun getLoaderName(declaration: KSClassDeclaration): String {
+    val fullName = getFullName(declaration).joinToString("")
+    return "${fullName}Loader"
+}
